# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.BatchDeleteFilesRequest:
    properties:
      file_ids:
        items:
          type: string
        type: array
    required:
    - file_ids
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.BroadcastMessageRequest:
    properties:
      message: {}
    required:
    - message
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateChannelRequest:
    properties:
      name:
        example: chat_room
        type: string
    required:
    - name
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.CreateStorageBucketRequest:
    properties:
      name:
        example: images
        type: string
      public:
        example: false
        type: boolean
    required:
    - name
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.TrackPresenceRequest:
    properties:
      data:
        additionalProperties: true
        type: object
      user_id:
        type: string
    required:
    - user_id
    type: object
...
  github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateFileMetadataRequest:
    properties:
      metadata:
        additionalProperties: true
        type: object
    required:
    - metadata
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the Sovrabase API server.
  title: Sovrabase API
  version: "1.0"
paths:
...
  /project/{id}/realtime:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: Both Websocket connection
      tags:
      - Realtime
  /project/{id}/realtime/channels:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: List Channels
      tags:
      - Realtime
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Channel
      tags:
      - Realtime
  /project/{id}/realtime/channels/{channel_id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel ID
        in: path
        name: channel_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Channel
      tags:
      - Realtime
  /project/{id}/realtime/channels/{channel_id}/broadcast:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel ID
        in: path
        name: channel_id
        required: true
        type: string
      - description: Broadcast message data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.BroadcastMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Broadcast Message
      tags:
      - Realtime
  /project/{id}/realtime/presence/{channel}:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel Name
        in: path
        name: channel
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Presence
      tags:
      - Realtime
  /project/{id}/realtime/presence/{channel}/track:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel Name
        in: path
        name: channel
        required: true
        type: string
      - description: Presence tracking data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.TrackPresenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Track Presence
      tags:
      - Realtime
  /project/{id}/realtime/presence/{channel}/untrack:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Channel Name
        in: path
        name: channel
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Untrack Presence
      tags:
      - Realtime
...
  /project/{id}/storage/buckets:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Storage Buckets
      tags:
      - Storage
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.CreateStorageBucketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Storage Bucket
      tags:
      - Storage
  /project/{id}/storage/buckets/{bucket_id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete Storage Bucket
      tags:
      - Storage
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Storage Bucket
      tags:
      - Storage
  /project/{id}/storage/buckets/{bucket_id}/files:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get Bucket Files
      tags:
      - Storage
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Upload File
      tags:
      - Storage
  /project/{id}/storage/buckets/{bucket_id}/files/{file_id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete File
      tags:
      - Storage
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get File
      tags:
      - Storage
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      - description: Metadata update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.UpdateFileMetadataRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Update metadata informations only !
      tags:
      - Storage
  /project/{id}/storage/buckets/{bucket_id}/files/{file_id}/info:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Get File Info
      tags:
      - Storage
  /project/{id}/storage/buckets/{bucket_id}/files/{file_id}/public-url:
    post:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Create Public URL
      tags:
      - Storage
  /project/{id}/storage/buckets/{bucket_id}/files/delete-batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Bucket ID
        in: path
        name: bucket_id
        required: true
        type: string
      - description: File IDs to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ketsuna-org_sovrabase_internal_models_requests.BatchDeleteFilesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - Bearer: []
      summary: Delete files in a batch operations !
      tags:
      - Storage
...
swagger: "2.0"