# internal/api/handlers/storage.go
package handlers

import (
	"net/http"

	"github.com/ketsuna-org/sovrabase/internal/models/requests"
)

// GetStorageBucketsHandler gets all storage buckets
// @Summary Get Storage Buckets
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Success 200
// @Router /project/{id}/storage/buckets [get]
func GetStorageBucketsHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get buckets logic
	w.WriteHeader(http.StatusOK)
}

// CreateStorageBucketHandler creates a new storage bucket
// @Summary Create Storage Bucket
// @Tags Storage
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param request body requests.CreateStorageBucketRequest true "Bucket creation data"
// @Success 200
// @Router /project/{id}/storage/buckets [post]
func CreateStorageBucketHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.CreateStorageBucketRequest
	_ = req
	// TODO: Implement create bucket logic
	w.WriteHeader(http.StatusOK)
}

// GetStorageBucketHandler gets a specific storage bucket
// @Summary Get Storage Bucket
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id} [get]
func GetStorageBucketHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get bucket logic
	w.WriteHeader(http.StatusOK)
}

// DeleteStorageBucketHandler deletes a storage bucket
// @Summary Delete Storage Bucket
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id} [delete]
func DeleteStorageBucketHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement delete bucket logic
	w.WriteHeader(http.StatusOK)
}

// GetBucketFilesHandler gets files in a bucket
// @Summary Get Bucket Files
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files [get]
func GetBucketFilesHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get files logic
	w.WriteHeader(http.StatusOK)
}

// UploadFileHandler uploads a file to a bucket
// @Summary Upload File
// @Tags Storage
// @Security Bearer
// @Accept multipart/form-data
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files [post]
func UploadFileHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement file upload logic
	w.WriteHeader(http.StatusOK)
}

// GetFileHandler gets a specific file
// @Summary Get File
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Param file_id path string true "File ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files/{file_id} [get]
func GetFileHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get file logic
	w.WriteHeader(http.StatusOK)
}

// DeleteFileHandler deletes a file
// @Summary Delete File
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Param file_id path string true "File ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files/{file_id} [delete]
func DeleteFileHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement delete file logic
	w.WriteHeader(http.StatusOK)
}

// GetFileInfoHandler gets file information
// @Summary Get File Info
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Param file_id path string true "File ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files/{file_id}/info [get]
func GetFileInfoHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get file info logic
	w.WriteHeader(http.StatusOK)
}

// CreatePublicURLHandler creates a public URL for a file
// @Summary Create Public URL
// @Tags Storage
// @Security Bearer
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Param file_id path string true "File ID"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files/{file_id}/public-url [post]
func CreatePublicURLHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement create public URL logic
	w.WriteHeader(http.StatusOK)
}

// BatchDeleteFilesHandler deletes files in a batch operation
// @Summary Delete files in a batch operations !
// @Tags Storage
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Param request body requests.BatchDeleteFilesRequest true "File IDs to delete"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files/delete-batch [post]
func BatchDeleteFilesHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.BatchDeleteFilesRequest
	_ = req
	// TODO: Implement batch delete files logic
	w.WriteHeader(http.StatusOK)
}

// UpdateFileMetadataHandler updates file metadata
// @Summary Update metadata informations only !
// @Tags Storage
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param bucket_id path string true "Bucket ID"
// @Param file_id path string true "File ID"
// @Param request body requests.UpdateFileMetadataRequest true "Metadata update data"
// @Success 200
// @Router /project/{id}/storage/buckets/{bucket_id}/files/{file_id} [patch]
func UpdateFileMetadataHandler(w http.ResponseWriter, r *http.Request) {
	var req requests.UpdateFileMetadataRequest
	_ = req
	// TODO: Implement update file metadata logic
	w.WriteHeader(http.StatusOK)
}